+++ src/modules/bin/mod.rs
@@ -15,6 +15,7 @@ mod console_module;
 mod map_module;
 // mod fn_module;
 // use fn_module::curry;
+mod boolean_module;
 mod fs_module;
 mod list_module;
 mod log_module;
@@ -49,6 +50,7 @@ pub fn get_module_map() -> HashMap<String, Expression> {
         String::from("List") => list_module::get(),
         String::from("Sys") => sys_module::get(),
         String::from("Into") => into_module::get(),
+        String::from("Boolean") => boolean_module::get(),
 
         // console control
         // Shell control